Implement PTY backend refactor foundation
