+++ src-tauri/src/pty/local_backend.rs
@@ -1,0 +1,535 @@
+use std::collections::HashMap;
+use std::io::{Read, Write};
+use std::sync::{Arc, Mutex};
+
+use portable_pty::{native_pty_system, CommandBuilder, MasterPty, PtySize};
+use tauri::AppHandle;
+use uuid::Uuid;
+
+use crate::errors::AppError;
+use crate::pty::backend::PtyBackend;
+use crate::pty::git_watcher::GitWatcher;
+use crate::pty::idle_detector::IdleDetector;
+use crate::pty::output::{PtyScreenSnapshot, SessionIoFanout};
+use crate::resource_monitor::ResourceLaunchRegistration;
+use crate::session::profile::IdleTuning;
+use crate::telegram::manager::OutputSenderMap;
+
+struct PtyInstance {
+    master: Arc<Mutex<Box<dyn MasterPty + Send>>>,
+    writer: Arc<Mutex<Box<dyn Write + Send>>>,
+    child: Option<Box<dyn portable_pty::Child + Send + Sync>>,
+    job: Option<crate::pty::job::JobObject>,
+}
+
+#[cfg(windows)]
+struct GitGuardEnv {
+    path: String,
+    pathext: String,
+    real_git: String,
+}
+
+#[cfg(windows)]
+fn resolve_real_git_path() -> Option<String> {
+    use std::os::windows::process::CommandExt;
+
+    const CREATE_NO_WINDOW: u32 = 0x08000000;
+    let mut cmd = std::process::Command::new("where.exe");
+    crate::pty::credentials::scrub_credentials_from_std_command(&mut cmd);
+    cmd.arg("git.exe").creation_flags(CREATE_NO_WINDOW);
+    let output = cmd.output().ok()?;
+    if !output.status.success() {
+        return None;
+    }
+
+    String::from_utf8_lossy(&output.stdout)
+        .lines()
+        .map(str::trim)
+        .find(|line| !line.is_empty())
+        .map(|line| line.to_string())
+}
+
+#[cfg(windows)]
+fn ensure_git_guard_wrapper() -> Result<std::path::PathBuf, AppError> {
+    let config_dir = crate::config::config_dir()
+        .ok_or_else(|| AppError::Other("Could not resolve app config directory".to_string()))?;
+    let guard_dir = config_dir.join("git-guard");
+    std::fs::create_dir_all(&guard_dir)
+        .map_err(|e| AppError::Other(format!("Failed to create git-guard dir: {}", e)))?;
+
+    let cmd_path = guard_dir.join("git.cmd");
+    let ps1_path = guard_dir.join("git-guard.ps1");
+
+    let cmd_content = "@echo off\r\npowershell.exe -NoProfile -ExecutionPolicy Bypass -File \"%~dp0git-guard.ps1\" %*\r\nexit /b %ERRORLEVEL%\r\n";
+    let ps1_content = r#"$ErrorActionPreference = 'Stop'
+$realGit = $env:AC_REAL_GIT
+if ([string]::IsNullOrWhiteSpace($realGit)) {
+  Write-Error 'AgentsCommander git guard: AC_REAL_GIT is not set.'
+  exit 1
+}
+
+$originalArgs = @($args)
+$target = (Get-Location).Path
+
+for ($i = 0; $i -lt $originalArgs.Count; $i++) {
+  $arg = [string]$originalArgs[$i]
+  if ($arg -eq '-C') {
+    if ($i + 1 -ge $originalArgs.Count) {
+      Write-Error 'AgentsCommander git guard: missing path after -C.'
+      exit 1
+    }
+
+    $next = [string]$originalArgs[$i + 1]
+    if ([System.IO.Path]::IsPathRooted($next)) {
+      $target = [System.IO.Path]::GetFullPath($next)
+    } else {
+      $target = [System.IO.Path]::GetFullPath((Join-Path -Path $target -ChildPath $next))
+    }
+    $i++
+    continue
+  }
+
+  if ($arg -eq '--git-dir' -or $arg -like '--git-dir=*' -or $arg -eq '--work-tree' -or $arg -like '--work-tree=*') {
+    Write-Error 'AgentsCommander git guard: --git-dir and --work-tree are not allowed in agent sessions.'
+    exit 1
+  }
+}
+
+function Test-AllowedGitTarget([string]$path) {
+  try {
+    $current = [System.IO.Path]::GetFullPath($path)
+  } catch {
+    return $false
+  }
+
+  while ($true) {
+    $name = [System.IO.Path]::GetFileName($current)
+    if ($name -like 'repo-*') {
+      return $true
+    }
+
+    $parent = Split-Path -Path $current -Parent
+    if ([string]::IsNullOrWhiteSpace($parent) -or $parent -eq $current) {
+      break
+    }
+    $current = $parent
+  }
+
+  return $false
+}
+
+if (-not (Test-AllowedGitTarget $target)) {
+  Write-Error ('AgentsCommander git guard: git is only allowed inside repo-* directories. Target path: ' + $target)
+  exit 1
+}
+
+& $realGit @originalArgs
+exit $LASTEXITCODE
+"#;
+
+    std::fs::write(&cmd_path, cmd_content)
+        .map_err(|e| AppError::Other(format!("Failed to write git.cmd guard: {}", e)))?;
+    std::fs::write(&ps1_path, ps1_content)
+        .map_err(|e| AppError::Other(format!("Failed to write git-guard.ps1: {}", e)))?;
+
+    Ok(guard_dir)
+}
+
+#[cfg(windows)]
+fn build_git_guard_env() -> Result<Option<GitGuardEnv>, AppError> {
+    let Some(real_git) = resolve_real_git_path() else {
+        log::warn!("[pty] git.exe not found; skipping PATH git guard wrapper");
+        return Ok(None);
+    };
+
+    let guard_dir = ensure_git_guard_wrapper()?;
+    let current_path = std::env::var_os("PATH").unwrap_or_default();
+    let mut path_entries: Vec<std::path::PathBuf> = vec![guard_dir];
+    path_entries.extend(std::env::split_paths(&current_path));
+    let path = std::env::join_paths(path_entries.iter())
+        .map_err(|e| AppError::Other(format!("Failed to join PATH for git guard: {}", e)))?
+        .to_string_lossy()
+        .to_string();
+
+    Ok(Some(GitGuardEnv {
+        path,
+        pathext: ".CMD;.BAT;.COM;.EXE".to_string(),
+        real_git,
+    }))
+}
+
+pub struct LocalProcessBackend {
+    ptys: Arc<Mutex<HashMap<Uuid, PtyInstance>>>,
+    fanout: SessionIoFanout,
+    git_watcher: Arc<GitWatcher>,
+}
+
+impl LocalProcessBackend {
+    pub fn new(
+        output_senders: OutputSenderMap,
+        idle_detector: Arc<IdleDetector>,
+        git_watcher: Arc<GitWatcher>,
+        ws_broadcaster: Option<crate::web::broadcast::WsBroadcaster>,
+    ) -> Self {
+        Self {
+            ptys: Arc::new(Mutex::new(HashMap::new())),
+            fanout: SessionIoFanout::new(output_senders, idle_detector, ws_broadcaster),
+            git_watcher,
+        }
+    }
+
+    #[allow(clippy::too_many_arguments)]
+    pub fn spawn<R: tauri::Runtime>(
+        &self,
+        id: Uuid,
+        cmd: &str,
+        args: &[String],
+        cwd: &str,
+        cols: u16,
+        rows: u16,
+        configured_env: &[(String, String)],
+        env_remove_keys: &[String],
+        extra_env: &[(String, String)],
+        idle_tuning: IdleTuning,
+        app_handle: AppHandle<R>,
+        mut resource_registration: Option<ResourceLaunchRegistration>,
+    ) -> Result<(), AppError> {
+        let pty_system = native_pty_system();
+        let spawn_cwd = crate::path_utils::normalize_windows_verbatim_path(cwd);
+
+        let size = PtySize {
+            rows,
+            cols,
+            pixel_width: 0,
+            pixel_height: 0,
+        };
+
+        let pair = pty_system
+            .openpty(size)
+            .map_err(|e| AppError::PtyError(e.to_string()))?;
+
+        let is_direct_exe = cmd.to_lowercase().ends_with(".exe")
+            || std::path::Path::new(cmd)
+                .extension()
+                .is_some_and(|ext| ext.eq_ignore_ascii_case("exe"));
+
+        let mut command = if cfg!(windows) && !is_direct_exe {
+            let mut c = CommandBuilder::new("cmd.exe");
+            c.arg("/C");
+            c.arg(cmd);
+            for arg in args {
+                c.arg(arg);
+            }
+            c
+        } else {
+            let mut c = CommandBuilder::new(cmd);
+            for arg in args {
+                c.arg(arg);
+            }
+            c
+        };
+        command.cwd(&spawn_cwd);
+        for key in env_remove_keys {
+            command.env_remove(key);
+        }
+        for (key, value) in configured_env {
+            command.env(key, value);
+        }
+        if !configured_env.is_empty() || !env_remove_keys.is_empty() {
+            log::info!(
+                "[pty] Applied {} configured env vars and removed {} inherited env vars for session {}",
+                configured_env.len(),
+                env_remove_keys.len(),
+                id
+            );
+        }
+        crate::pty::credentials::apply_credential_env_to_pty_command(&mut command, extra_env);
+        command.env("TERM", "xterm-256color");
+
+        if !extra_env.is_empty() {
+            log::info!(
+                "[pty] Applied {} per-process credential environment variables for session {}",
+                extra_env.len(),
+                id
+            );
+        }
+
+        if let Some(git_ceiling_dirs) =
+            crate::config::session_context::git_ceiling_directories_for_session_root(&spawn_cwd)
+        {
+            command.env("GIT_CEILING_DIRECTORIES", &git_ceiling_dirs);
+            log::info!(
+                "[pty] Applied GIT_CEILING_DIRECTORIES for session cwd {}: {}",
+                spawn_cwd,
+                git_ceiling_dirs
+            );
+
+            #[cfg(windows)]
+            if let Some(git_guard_env) = build_git_guard_env()? {
+                command.env("PATH", &git_guard_env.path);
+                command.env("PATHEXT", &git_guard_env.pathext);
+                command.env("AC_REAL_GIT", &git_guard_env.real_git);
+                log::info!(
+                    "[pty] Enabled git guard wrapper for session cwd {}",
+                    spawn_cwd
+                );
+            }
+        }
+
+        let mut child = pair
+            .slave
+            .spawn_command(command)
+            .map_err(|e| AppError::PtyError(e.to_string()))?;
+        log::info!(
+            "[pty] Spawned session {} with child pid {:?}",
+            id,
+            child.process_id()
+        );
+
+        let job = child
+            .process_id()
+            .and_then(crate::pty::job::JobObject::for_child);
+
+        if let Some(registration) = resource_registration.as_mut() {
+            let Some(pid) = child.process_id() else {
+                let _ = child.kill();
+                return Err(AppError::PtyError(
+                    "Resource Monitor could not capture spawned child pid".to_string(),
+                ));
+            };
+            if let Err(err) = registration.register_root_pid(pid) {
+                let _ = child.kill();
+                return Err(AppError::PtyError(err));
+            }
+        }
+
+        drop(pair.slave);
+
+        let writer = match pair.master.take_writer() {
+            Ok(writer) => writer,
+            Err(e) => {
+                if let Some(registration) = resource_registration.as_ref() {
+                    registration.rollback_registered();
+                }
+                return Err(AppError::PtyError(e.to_string()));
+            }
+        };
+
+        let mut reader = match pair.master.try_clone_reader() {
+            Ok(reader) => reader,
+            Err(e) => {
+                if let Some(registration) = resource_registration.as_ref() {
+                    registration.rollback_registered();
+                }
+                return Err(AppError::PtyError(e.to_string()));
+            }
+        };
+
+        let instance = PtyInstance {
+            master: Arc::new(Mutex::new(pair.master)),
+            writer: Arc::new(Mutex::new(writer)),
+            child: Some(child),
+            job,
+        };
+
+        self.ptys.lock().unwrap().insert(id, instance);
+        self.fanout.register_session(id, idle_tuning, rows, cols);
+
+        let session_id_str = id.to_string();
+        let fanout = self.fanout.clone();
+        std::thread::spawn(move || {
+            let mut buf = [0u8; 4096];
+            loop {
+                match reader.read(&mut buf) {
+                    Ok(0) => break,
+                    Ok(n) => {
+                        fanout.handle_output(&app_handle, id, &session_id_str, buf[..n].to_vec())
+                    }
+                    Err(_) => break,
+                }
+            }
+        });
+
+        Ok(())
+    }
+}
+
+impl PtyBackend for LocalProcessBackend {
+    fn as_any(&self) -> &dyn std::any::Any {
+        self
+    }
+
+    fn write(&self, id: Uuid, data: &[u8]) -> Result<(), AppError> {
+        let ptys = self.ptys.lock().unwrap();
+        let instance = ptys
+            .get(&id)
+            .ok_or_else(|| AppError::SessionNotFound(id.to_string()))?;
+
+        let mut writer = instance.writer.lock().unwrap();
+        writer
+            .write_all(data)
+            .map_err(|e| AppError::PtyError(e.to_string()))?;
+        writer
+            .flush()
+            .map_err(|e| AppError::PtyError(e.to_string()))?;
+
+        Ok(())
+    }
+
+    fn has_session(&self, id: Uuid) -> bool {
+        self.ptys.lock().unwrap().contains_key(&id)
+    }
+
+    fn resize(&self, id: Uuid, cols: u16, rows: u16) -> Result<(), AppError> {
+        self.fanout.record_resize(id);
+
+        let ptys = self.ptys.lock().unwrap();
+        let instance = ptys
+            .get(&id)
+            .ok_or_else(|| AppError::SessionNotFound(id.to_string()))?;
+
+        let master = instance.master.lock().unwrap();
+        master
+            .resize(PtySize {
+                rows,
+                cols,
+                pixel_width: 0,
+                pixel_height: 0,
+            })
+            .map_err(|e| AppError::PtyError(e.to_string()))?;
+
+        self.fanout.resize_screen_and_broadcast(id, cols, rows);
+
+        Ok(())
+    }
+
+    fn kill(&self, id: Uuid) -> Result<(), AppError> {
+        let instance = {
+            let mut ptys = self.ptys.lock().unwrap();
+            ptys.remove(&id)
+        };
+
+        if let Some(mut instance) = instance {
+            if let Some(job) = instance.job.take() {
+                job.terminate();
+            }
+            if let Some(mut child) = instance.child.take() {
+                let pid = child.process_id();
+                match child.try_wait() {
+                    Ok(Some(status)) => {
+                        log::info!(
+                            "[pty] Session {} child pid {:?} already exited: {:?}",
+                            id,
+                            pid,
+                            status
+                        );
+                    }
+                    Ok(None) => {
+                        if let Err(e) = child.kill() {
+                            log::warn!(
+                                "[pty] Failed to kill session {} child pid {:?}: {}",
+                                id,
+                                pid,
+                                e
+                            );
+                        }
+                        reap_child_in_background(id, pid, child);
+                    }
+                    Err(e) => {
+                        log::warn!(
+                            "[pty] Failed to poll session {} child pid {:?}: {}",
+                            id,
+                            pid,
+                            e
+                        );
+                        if let Err(kill_err) = child.kill() {
+                            log::warn!(
+                                "[pty] Failed to kill session {} child pid {:?} after poll error: {}",
+                                id,
+                                pid,
+                                kill_err
+                            );
+                        }
+                        reap_child_in_background(id, pid, child);
+                    }
+                }
+            }
+        }
+
+        self.fanout.remove_session(id);
+        self.git_watcher.remove_session(id);
+
+        Ok(())
+    }
+
+    fn terminate_job_for_session(&self, id: Uuid) -> bool {
+        let ptys = self.ptys.lock().unwrap();
+        match ptys.get(&id).and_then(|inst| inst.job.as_ref()) {
+            Some(job) => {
+                job.terminate();
+                true
+            }
+            None => false,
+        }
+    }
+
+    fn kill_all_jobs(&self) -> (usize, usize) {
+        let ptys = self.ptys.lock().unwrap_or_else(|e| e.into_inner());
+        let mut terminated = 0;
+        let mut jobless = 0;
+        for (id, instance) in ptys.iter() {
+            match instance.job.as_ref() {
+                Some(job) => {
+                    job.terminate();
+                    terminated += 1;
+                    log::info!("[pty] terminated job object for session {id} at shutdown");
+                }
+                None => jobless += 1,
+            }
+        }
+        (terminated, jobless)
+    }
+
+    fn get_screen_snapshot(&self, id: Uuid) -> Option<PtyScreenSnapshot> {
+        self.fanout.get_screen_snapshot(id)
+    }
+
+    fn get_pty_size(&self, id: Uuid) -> Option<(u16, u16)> {
+        self.fanout.get_pty_size(id)
+    }
+
+    fn register_response_watcher(
+        &self,
+        session_id: Uuid,
+        request_id: String,
+        response_dir: std::path::PathBuf,
+    ) {
+        self.fanout
+            .register_response_watcher(session_id, request_id, response_dir);
+    }
+}
+
+fn reap_child_in_background(
+    session_id: Uuid,
+    pid: Option<u32>,
+    mut child: Box<dyn portable_pty::Child + Send + Sync>,
+) {
+    std::thread::spawn(move || match child.wait() {
+        Ok(status) => {
+            log::info!(
+                "[pty] Reaped session {} child pid {:?}: {:?}",
+                session_id,
+                pid,
+                status
+            );
+        }
+        Err(e) => {
+            log::warn!(
+                "[pty] Failed to reap session {} child pid {:?}: {}",
+                session_id,
+                pid,
+                e
+            );
+        }
+    });
+}
