+++ src-tauri/src/pty/mod.rs
@@ -1,6 +1,9 @@
+pub mod backend;
 pub mod credentials;
 pub mod git_watcher;
 pub mod idle_detector;
 pub mod inject;
 pub mod job; // #632 - per-agent Job Object for tree-kill
+pub mod local_backend;
 pub mod manager;
+pub mod output;
