+++ src-tauri/src/pty/output.rs
@@ -1,0 +1,394 @@
+use std::collections::HashMap;
+use std::sync::{Arc, Mutex};
+
+use tauri::{AppHandle, Emitter};
+use uuid::Uuid;
+
+use crate::pty::idle_detector::IdleDetector;
+use crate::session::profile::IdleTuning;
+use crate::telegram::manager::OutputSenderMap;
+
+/// Tracks active response marker watchers per session.
+/// Key: (session_id, request_id) -> accumulated output buffer.
+/// The read loop scans for %%AC_RESPONSE::<rid>::START/END%% markers.
+pub type ResponseWatcherMap = Arc<Mutex<HashMap<(Uuid, String), ResponseWatcher>>>;
+
+pub struct ResponseWatcher {
+    pub response_dir: std::path::PathBuf,
+    pub buffer: Option<String>,
+    pub capturing: bool,
+}
+
+#[derive(Clone)]
+pub struct SessionIoFanout {
+    output_senders: OutputSenderMap,
+    idle_detector: Arc<IdleDetector>,
+    response_watchers: ResponseWatcherMap,
+    ws_broadcaster: Option<crate::web::broadcast::WsBroadcaster>,
+    screen_parsers: Arc<Mutex<HashMap<Uuid, ScreenReplayState>>>,
+}
+
+pub struct PtyScreenSnapshot {
+    pub data: Vec<u8>,
+    pub rows: u16,
+    pub cols: u16,
+    pub sequence: u64,
+}
+
+struct ScreenReplayState {
+    parser: vt100::Parser,
+    output_sequence: u64,
+}
+
+#[derive(Clone, serde::Serialize)]
+#[serde(rename_all = "camelCase")]
+struct PtyOutputPayload {
+    session_id: String,
+    data: Vec<u8>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    sequence: Option<u64>,
+}
+
+impl SessionIoFanout {
+    pub fn new(
+        output_senders: OutputSenderMap,
+        idle_detector: Arc<IdleDetector>,
+        ws_broadcaster: Option<crate::web::broadcast::WsBroadcaster>,
+    ) -> Self {
+        Self {
+            output_senders,
+            idle_detector,
+            response_watchers: Arc::new(Mutex::new(HashMap::new())),
+            ws_broadcaster,
+            screen_parsers: Arc::new(Mutex::new(HashMap::new())),
+        }
+    }
+
+    pub fn register_session(&self, id: Uuid, idle_tuning: IdleTuning, rows: u16, cols: u16) {
+        self.idle_detector.register_session(id, idle_tuning);
+        let replay = ScreenReplayState {
+            parser: vt100::Parser::new(rows, cols, 0),
+            output_sequence: 0,
+        };
+        self.screen_parsers.lock().unwrap().insert(id, replay);
+    }
+
+    pub fn handle_output<R: tauri::Runtime>(
+        &self,
+        app_handle: &AppHandle<R>,
+        id: Uuid,
+        session_id_str: &str,
+        data: Vec<u8>,
+    ) {
+        let n = data.len();
+
+        self.idle_detector.touch_silence(id);
+
+        let text = String::from_utf8_lossy(&data);
+        if text.contains('\u{FFFD}') {
+            log::debug!(
+                "[PTY] session {} chunk had invalid UTF-8 at buffer boundary ({} bytes, {} replacement chars)",
+                id,
+                n,
+                text.matches('\u{FFFD}').count()
+            );
+        }
+
+        if output_has_printable_activity(&text) {
+            self.idle_detector.record_activity_with_bytes(id, n);
+        } else {
+            log::trace!(
+                "[idle] SKIPPED activity for {} ({} bytes, escape-only output)",
+                &id.to_string()[..8],
+                n
+            );
+        }
+
+        scan_response_markers(id, &text, &self.response_watchers);
+
+        if let Ok(senders) = self.output_senders.lock() {
+            if let Some(tx) = senders.get(&id) {
+                let _ = tx.try_send(data.clone());
+            }
+        }
+
+        let sequence = if let Ok(mut parsers) = self.screen_parsers.lock() {
+            if let Some(state) = parsers.get_mut(&id) {
+                state.parser.process(&data);
+                state.output_sequence = state.output_sequence.saturating_add(1);
+                Some(state.output_sequence)
+            } else {
+                None
+            }
+        } else {
+            None
+        };
+
+        if let Some(ref bc) = self.ws_broadcaster {
+            bc.broadcast_pty_output(session_id_str, &data);
+        }
+
+        let payload = PtyOutputPayload {
+            session_id: session_id_str.to_string(),
+            data,
+            sequence,
+        };
+        let _ = app_handle.emit("pty_output", payload);
+    }
+
+    pub fn record_resize(&self, id: Uuid) {
+        self.idle_detector.record_resize(id);
+    }
+
+    pub fn resize_screen_and_broadcast(&self, id: Uuid, cols: u16, rows: u16) {
+        if let Ok(mut parsers) = self.screen_parsers.lock() {
+            if let Some(state) = parsers.get_mut(&id) {
+                state.parser.set_size(rows, cols);
+            }
+        }
+
+        if let Some(ref bc) = self.ws_broadcaster {
+            bc.broadcast_event(
+                "pty_resized",
+                &serde_json::json!({
+                    "sessionId": id.to_string(),
+                    "cols": cols,
+                    "rows": rows,
+                }),
+            );
+        }
+    }
+
+    pub fn remove_session(&self, id: Uuid) {
+        self.idle_detector.remove_session(id);
+
+        if let Ok(mut watchers) = self.response_watchers.lock() {
+            watchers.retain(|(sid, _), _| *sid != id);
+        }
+
+        if let Ok(mut parsers) = self.screen_parsers.lock() {
+            parsers.remove(&id);
+        }
+    }
+
+    pub fn get_screen_snapshot(&self, id: Uuid) -> Option<PtyScreenSnapshot> {
+        let parsers = self.screen_parsers.lock().ok()?;
+        let state = parsers.get(&id)?;
+        let screen = state.parser.screen();
+        let (rows, cols) = screen.size();
+        Some(PtyScreenSnapshot {
+            data: screen.contents_formatted(),
+            rows,
+            cols,
+            sequence: state.output_sequence,
+        })
+    }
+
+    pub fn get_pty_size(&self, id: Uuid) -> Option<(u16, u16)> {
+        let parsers = self.screen_parsers.lock().ok()?;
+        let state = parsers.get(&id)?;
+        Some(state.parser.screen().size())
+    }
+
+    pub fn register_response_watcher(
+        &self,
+        session_id: Uuid,
+        request_id: String,
+        response_dir: std::path::PathBuf,
+    ) {
+        if let Ok(mut watchers) = self.response_watchers.lock() {
+            watchers.insert(
+                (session_id, request_id),
+                ResponseWatcher {
+                    response_dir,
+                    buffer: None,
+                    capturing: false,
+                },
+            );
+        }
+    }
+}
+
+/// Strip ANSI escape sequences so marker detection ignores color, cursor,
+/// title, hyperlink, shell-integration, and device-control noise.
+pub(crate) fn strip_ansi_csi(s: &str) -> String {
+    let mut out = String::with_capacity(s.len());
+    let mut chars = s.chars().peekable();
+    while let Some(c) = chars.next() {
+        if c == '\x1b' {
+            match chars.peek() {
+                Some(&'[') => {
+                    chars.next();
+                    while let Some(&next) = chars.peek() {
+                        chars.next();
+                        if ('@'..='~').contains(&next) {
+                            break;
+                        }
+                    }
+                }
+                Some(&']') => {
+                    chars.next();
+                    while let Some(&ch) = chars.peek() {
+                        if ch == '\x07' {
+                            chars.next();
+                            break;
+                        }
+                        if ch == '\x1b' {
+                            chars.next();
+                            if chars.peek() == Some(&'\\') {
+                                chars.next();
+                                break;
+                            }
+                            continue;
+                        }
+                        chars.next();
+                    }
+                }
+                Some(&'P') => {
+                    chars.next();
+                    while let Some(&ch) = chars.peek() {
+                        if ch == '\x1b' {
+                            chars.next();
+                            if chars.peek() == Some(&'\\') {
+                                chars.next();
+                                break;
+                            }
+                            continue;
+                        }
+                        chars.next();
+                    }
+                }
+                Some(_) => {
+                    chars.next();
+                }
+                None => {}
+            }
+        } else {
+            out.push(c);
+        }
+    }
+    out
+}
+
+pub(crate) fn output_has_printable_activity(text: &str) -> bool {
+    let is_printable = |c: char| c > ' ' && c != '\u{FFFD}';
+    if text.contains('\x1b') {
+        strip_ansi_csi(text).chars().any(is_printable)
+    } else {
+        text.chars().any(is_printable)
+    }
+}
+
+/// Scan PTY output text for %%AC_RESPONSE::<rid>::START/END%% markers.
+pub(crate) fn scan_response_markers(session_id: Uuid, text: &str, watchers: &ResponseWatcherMap) {
+    let Ok(mut watchers) = watchers.lock() else {
+        return;
+    };
+
+    let keys: Vec<(Uuid, String)> = watchers
+        .keys()
+        .filter(|(sid, _)| *sid == session_id)
+        .cloned()
+        .collect();
+
+    for key in keys {
+        let (_, ref rid) = key;
+        let start_marker = format!("%%AC_RESPONSE::{}::START%%", rid);
+        let end_marker = format!("%%AC_RESPONSE::{}::END%%", rid);
+
+        let watcher = match watchers.get_mut(&key) {
+            Some(w) => w,
+            None => continue,
+        };
+
+        if watcher.capturing {
+            if let Some(end_pos) = text.find(&end_marker) {
+                let chunk = &text[..end_pos];
+                if let Some(ref mut buf) = watcher.buffer {
+                    buf.push_str(chunk);
+                }
+
+                let response_content = watcher.buffer.take().unwrap_or_default().trim().to_string();
+                write_response_file(&watcher.response_dir, rid, response_content);
+                watchers.remove(&key);
+                return;
+            } else if let Some(ref mut buf) = watcher.buffer {
+                buf.push_str(text);
+            }
+        } else if let Some(start_pos) = text.find(&start_marker) {
+            watcher.capturing = true;
+            let after_start = &text[start_pos + start_marker.len()..];
+
+            if let Some(end_pos) = after_start.find(&end_marker) {
+                let content = after_start[..end_pos].trim().to_string();
+                write_response_file(&watcher.response_dir, rid, content);
+                watchers.remove(&key);
+                return;
+            } else {
+                watcher.buffer = Some(after_start.to_string());
+            }
+        }
+    }
+}
+
+fn write_response_file(response_dir: &std::path::Path, request_id: &str, content: String) {
+    let response_path = response_dir.join(format!("{}.json", request_id));
+    if let Err(e) = std::fs::create_dir_all(response_dir) {
+        log::warn!("Failed to create responses dir: {}", e);
+    }
+
+    let response_json = serde_json::json!({
+        "requestId": request_id,
+        "content": content,
+        "timestamp": chrono::Utc::now().to_rfc3339(),
+    });
+
+    match serde_json::to_string_pretty(&response_json) {
+        Ok(json) => {
+            if let Err(e) = std::fs::write(&response_path, json) {
+                log::warn!("Failed to write response file: {}", e);
+            } else {
+                log::info!("Captured response for request {}", request_id);
+            }
+        }
+        Err(e) => log::warn!("Failed to serialize response: {}", e),
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn printable_activity_ignores_ansi_only_chunks() {
+        assert!(!output_has_printable_activity("\x1b[31m\x1b[0m"));
+        assert!(!output_has_printable_activity("\x1b]0;title\x07"));
+        assert!(output_has_printable_activity("\x1b[31mready\x1b[0m"));
+    }
+
+    #[test]
+    fn response_marker_capture_writes_trimmed_json() {
+        let dir = tempfile::tempdir().expect("tempdir");
+        let session_id = Uuid::new_v4();
+        let watchers: ResponseWatcherMap = Arc::new(Mutex::new(HashMap::new()));
+        watchers.lock().unwrap().insert(
+            (session_id, "r1".to_string()),
+            ResponseWatcher {
+                response_dir: dir.path().to_path_buf(),
+                buffer: None,
+                capturing: false,
+            },
+        );
+
+        scan_response_markers(
+            session_id,
+            "before %%AC_RESPONSE::r1::START%% {\"ok\": true} %%AC_RESPONSE::r1::END%% after",
+            &watchers,
+        );
+
+        let json = std::fs::read_to_string(dir.path().join("r1.json")).expect("response json");
+        assert!(json.contains("\"requestId\": \"r1\""));
+        assert!(json.contains("\"content\": \"{\\\"ok\\\": true}\""));
+        assert!(watchers.lock().unwrap().is_empty());
+    }
+}
