+++ src-tauri/src/session/manager.rs
@@ -10,6 +10,7 @@ use super::session::{
 };
 use crate::config::settings::WindowGeometry;
 use crate::errors::AppError;
+use crate::pty::backend::SessionBackendKind;
 
 #[derive(Clone)]
 pub struct SessionManager {
@@ -47,6 +48,7 @@ impl SessionManager {
         agent_label: Option<String>,
         git_repos: Vec<SessionRepo>,
         is_coordinator: bool,
+        backend_kind: SessionBackendKind,
     ) -> Result<Session, AppError> {
         let id = Uuid::new_v4();
 
@@ -59,6 +61,7 @@ impl SessionManager {
             name,
             shell,
             shell_args,
+            backend_kind,
             effective_shell_args: None,
             created_at: chrono::Utc::now(),
             working_directory,
@@ -806,6 +809,7 @@ mod tests {
                 None,
                 Vec::new(),
                 false,
+                crate::pty::backend::SessionBackendKind::LocalProcess,
             )
             .await
             .expect("create_session should succeed");
@@ -849,6 +853,7 @@ mod tests {
                 None,
                 Vec::new(),
                 false,
+                crate::pty::backend::SessionBackendKind::LocalProcess,
             )
             .await
             .expect("create_session should succeed");
@@ -905,6 +910,7 @@ mod tests {
                 None,
                 Vec::new(),
                 false,
+                crate::pty::backend::SessionBackendKind::LocalProcess,
             )
             .await
             .expect("create_session should succeed");
@@ -948,6 +954,7 @@ mod tests {
                 None,
                 Vec::new(),
                 true,
+                crate::pty::backend::SessionBackendKind::LocalProcess,
             )
             .await
             .expect("create_session should succeed");
@@ -982,6 +989,7 @@ mod tests {
                 None,
                 Vec::new(),
                 true,
+                crate::pty::backend::SessionBackendKind::LocalProcess,
             )
             .await
             .expect("create_session should succeed");
@@ -1012,6 +1020,7 @@ mod tests {
                 None,
                 Vec::new(),
                 false,
+                crate::pty::backend::SessionBackendKind::LocalProcess,
             )
             .await
             .expect("create_session should succeed");
@@ -1040,6 +1049,7 @@ mod tests {
                 None,
                 Vec::new(),
                 true,
+                crate::pty::backend::SessionBackendKind::LocalProcess,
             )
             .await
             .expect("create_session should succeed");
@@ -1069,6 +1079,7 @@ mod tests {
                 None,
                 Vec::new(),
                 true,
+                crate::pty::backend::SessionBackendKind::LocalProcess,
             )
             .await
             .expect("create_session should succeed");
@@ -1108,6 +1119,7 @@ mod tests {
                 None,
                 Vec::new(),
                 true,
+                crate::pty::backend::SessionBackendKind::LocalProcess,
             )
             .await
             .expect("create_session should succeed");
@@ -1135,6 +1147,7 @@ mod tests {
                 None,
                 Vec::new(),
                 true,
+                crate::pty::backend::SessionBackendKind::LocalProcess,
             )
             .await
             .expect("create_session should succeed");
@@ -1163,6 +1176,7 @@ mod tests {
                 None,
                 Vec::new(),
                 true,
+                crate::pty::backend::SessionBackendKind::LocalProcess,
             )
             .await
             .expect("create_session should succeed");
@@ -1212,6 +1226,7 @@ mod tests {
                 None,
                 Vec::new(),
                 false,
+                crate::pty::backend::SessionBackendKind::LocalProcess,
             )
             .await
             .expect("create_session should succeed");
@@ -1236,6 +1251,7 @@ mod tests {
                 None,
                 Vec::new(),
                 true,
+                crate::pty::backend::SessionBackendKind::LocalProcess,
             )
             .await
             .expect("create_session should succeed");
@@ -1270,6 +1286,7 @@ mod tests {
                 None,
                 Vec::new(),
                 true,
+                crate::pty::backend::SessionBackendKind::LocalProcess,
             )
             .await
             .expect("create_session should succeed");
@@ -1319,6 +1336,7 @@ mod tests {
                 None,
                 Vec::new(),
                 true,
+                crate::pty::backend::SessionBackendKind::LocalProcess,
             )
             .await
             .expect("create_session should succeed");
@@ -1352,6 +1370,7 @@ mod tests {
                 None,
                 Vec::new(),
                 false,
+                crate::pty::backend::SessionBackendKind::LocalProcess,
             )
             .await
             .expect("create_session should succeed");
@@ -1388,6 +1407,7 @@ mod tests {
                 None,
                 Vec::new(),
                 false,
+                crate::pty::backend::SessionBackendKind::LocalProcess,
             )
             .await
             .expect("create_session should succeed");
@@ -1419,6 +1439,7 @@ mod tests {
                 None,
                 Vec::new(),
                 false,
+                crate::pty::backend::SessionBackendKind::LocalProcess,
             )
             .await
             .expect("create_session should succeed");
@@ -1443,6 +1464,7 @@ mod tests {
                 None,
                 Vec::new(),
                 false,
+                crate::pty::backend::SessionBackendKind::LocalProcess,
             )
             .await
             .expect("create first session");
@@ -1455,6 +1477,7 @@ mod tests {
                 None,
                 Vec::new(),
                 false,
+                crate::pty::backend::SessionBackendKind::LocalProcess,
             )
             .await
             .expect("create second session");
@@ -1485,6 +1508,7 @@ mod tests {
                 None,
                 vec![],
                 true,
+                crate::pty::backend::SessionBackendKind::LocalProcess,
             )
             .await
             .unwrap();
@@ -1515,6 +1539,7 @@ mod tests {
                 None,
                 vec![],
                 false,
+                crate::pty::backend::SessionBackendKind::LocalProcess,
             )
             .await
             .unwrap();
@@ -1533,6 +1558,7 @@ mod tests {
                 None,
                 vec![],
                 false,
+                crate::pty::backend::SessionBackendKind::LocalProcess,
             )
             .await
             .unwrap();
@@ -1576,6 +1602,7 @@ mod tests {
                 Some("Architect".into()),
                 vec![],
                 true, // is_coordinator
+                crate::pty::backend::SessionBackendKind::LocalProcess,
             )
             .await
             .unwrap();
@@ -1617,6 +1644,7 @@ mod tests {
                 None,
                 vec![],
                 false,
+                crate::pty::backend::SessionBackendKind::LocalProcess,
             )
             .await
             .unwrap();
@@ -1657,6 +1685,7 @@ mod tests {
                 None,
                 vec![],
                 true, // is_coordinator
+                crate::pty::backend::SessionBackendKind::LocalProcess,
             )
             .await
             .unwrap();
@@ -1669,6 +1698,7 @@ mod tests {
                 None,
                 vec![],
                 false,
+                crate::pty::backend::SessionBackendKind::LocalProcess,
             )
             .await
             .unwrap();
@@ -1693,6 +1723,7 @@ mod tests {
                 None,
                 vec![],
                 true,
+                crate::pty::backend::SessionBackendKind::LocalProcess,
             )
             .await
             .unwrap();
@@ -1706,6 +1737,7 @@ mod tests {
                 None,
                 vec![],
                 false,
+                crate::pty::backend::SessionBackendKind::LocalProcess,
             )
             .await
             .unwrap();
@@ -1719,6 +1751,7 @@ mod tests {
                 None,
                 vec![],
                 false,
+                crate::pty::backend::SessionBackendKind::LocalProcess,
             )
             .await
             .unwrap();
@@ -1743,6 +1776,7 @@ mod tests {
                 None,
                 vec![],
                 true,
+                crate::pty::backend::SessionBackendKind::LocalProcess,
             )
             .await
             .unwrap();
@@ -1756,6 +1790,7 @@ mod tests {
                 None,
                 vec![],
                 false,
+                crate::pty::backend::SessionBackendKind::LocalProcess,
             )
             .await
             .unwrap();
@@ -1780,6 +1815,7 @@ mod tests {
                 None,
                 vec![],
                 true,
+                crate::pty::backend::SessionBackendKind::LocalProcess,
             )
             .await
             .unwrap();
@@ -1793,6 +1829,7 @@ mod tests {
                 None,
                 vec![],
                 false,
+                crate::pty::backend::SessionBackendKind::LocalProcess,
             )
             .await
             .unwrap();
